Update config.go, size_enforcer.go, and 5 more files...
