+++ cli/kubernetes/provider.go
@@ -488,10 +488,10 @@ func (provider *Provider) handleRemovalError(err error) error {
 	return err
 }
 
-func (provider *Provider) CreateConfigMap(ctx context.Context, namespace string, configMapName string, serializedRulePolicy string, serializedContract string, serializedMizuConfig string) error {
+func (provider *Provider) CreateConfigMap(ctx context.Context, namespace string, configMapName string, serializedValidationRules string, serializedContract string, serializedMizuConfig string) error {
 	configMapData := make(map[string]string, 0)
-	if serializedRulePolicy != "" {
-		configMapData[shared.RulePolicyFileName] = serializedRulePolicy
+	if serializedValidationRules != "" {
+		configMapData[shared.ValidationRulesFileName] = serializedValidationRules
 	}
 	if serializedContract != "" {
 		configMapData[shared.ContractFileName] = serializedContract
